fix(onboard): extract agent resume cleanup
